+++ src/utils/item-types.js
@@ -1,16 +1,25 @@
-// The maps types available for a TCR item column to their lengths in bytes.
-// Items marked with null are not stored on-chain.
-
-export const ADDRESS = 'address'
-export const NUMBER = 'number'
-export const TEXT = 'text'
-export const LONGTEXT = 'longtext'
-export const BOOLEAN = 'boolean'
+const ADDRESS = 'address'
+const NUMBER = 'number'
+const TEXT = 'text'
+const BOOLEAN = 'boolean'
 
 export default {
-  [ADDRESS]: 20,
-  [NUMBER]: 32,
-  [TEXT]: 64,
-  [LONGTEXT]: null,
-  [BOOLEAN]: 1
+  ADDRESS,
+  NUMBER,
+  TEXT,
+  BOOLEAN
+}
+
+export const typeToSolidity = {
+  [ADDRESS]: 'address',
+  [NUMBER]: 'int256',
+  [TEXT]: 'string',
+  [BOOLEAN]: 'bool'
+}
+
+export const typeDefaultValues = {
+  [ADDRESS]: '',
+  [TEXT]: '',
+  [BOOLEAN]: false,
+  [NUMBER]: 0
 }